+++ common/layer_tag_manager.hpp
@@ -52,7 +52,7 @@ public:
 
 private:
 
-    typedef boost::bimap<std::string,int> StringBimap;
+    typedef boost::bimap<std::string,size_t> StringBimap;
     typedef StringBimap::value_type StringBimapItem;
     StringBimap m_;
